Proyecto_FinalE/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Caching==2.3.0
psycopg[binary]==3.2.9
redis==5.0.7
